+++ src/argilla/__init__.py
@@ -66,12 +66,16 @@ if _TYPE_CHECKING:
         read_datasets,
         read_pandas,
     )
-    from argilla.client.feedback import (
-        FeedbackDataset,
+    from argilla.client.feedback.dataset import FeedbackDataset
+    from argilla.client.feedback.schemas import (
         FeedbackRecord,
+        LabelQuestion,
+        MultiLabelQuestion,
         RatingQuestion,
+        ResponseSchema,
         TextField,
         TextQuestion,
+        ValueSchema,
     )
     from argilla.client.models import (
         Text2TextRecord,
@@ -121,12 +125,16 @@ _import_structure = {
         "read_datasets",
         "read_pandas",
     ],
-    "client.feedback": [
-        "FeedbackDataset",
+    "client.feedback.dataset": ["FeedbackDataset"],
+    "client.feedback.schemas": [
+        "FeedbackRecord",
+        "LabelQuestion",
+        "MultiLabelQuestion",
+        "RatingQuestion",
+        "ResponseSchema",
         "TextField",
         "TextQuestion",
-        "RatingQuestion",
-        "FeedbackRecord",
+        "ValueSchema",
     ],
     "client.workspaces": ["Workspace"],
     "monitoring.model_monitor": ["monitor"],