+++ src/argilla/server/alembic/env.py
@@ -15,6 +15,8 @@
 from logging.config import fileConfig
 
 from alembic import context
+from argilla.server.database import Base
+from argilla.server.models.models import *  # noqa
 from argilla.server.settings import settings
 from sqlalchemy import engine_from_config, pool
 
@@ -34,7 +36,7 @@ if config.config_file_name is not None:
 # for 'autogenerate' support
 # from myapp import mymodel
 # target_metadata = mymodel.Base.metadata
-target_metadata = None
+target_metadata = Base.metadata
 
 # other values from the config, defined by the needs of env.py,
 # can be acquired: