+++ src/argilla/server/database.py
@@ -14,6 +14,7 @@
 import os
 from collections import OrderedDict
 from sqlite3 import Connection as SQLite3Connection
+from typing import TYPE_CHECKING, Generator
 
 from sqlalchemy import create_engine, event
 from sqlalchemy.engine import Engine
@@ -21,6 +22,9 @@ from sqlalchemy.orm import DeclarativeBase, sessionmaker
 
 import argilla
 from argilla.server.settings import settings
+
+if TYPE_CHECKING:
+    from sqlalchemy.orm import Session
 
 ALEMBIC_CONFIG_FILE = os.path.normpath(os.path.join(os.path.dirname(argilla.__file__), "alembic.ini"))
 TAGGED_REVISIONS = OrderedDict({"1.7": "1769ee58fbb4", "1.8": "ae5522b4c674"})
@@ -38,7 +42,7 @@ engine = create_engine(settings.database_url)
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
 
-def get_db():
+def get_db() -> Generator["Session", None, None]:
     try:
         db = SessionLocal()
         yield db